refactor(ui-overlays): extract search and help renderers
